# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TeamPulse API",
        "description": "Документация API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Вход",
                "operationId": "a7997e7c1e3bfed64d4e56de82415931",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Access токен",
                                            "type": "string"
                                        },
                                        "expirationDate": {
                                            "description": "Дата экспирации токена",
                                            "type": "integer",
                                            "format": "timestamp"
                                        },
                                        "id": {
                                            "description": "ID пользователя",
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "role": {
                                            "description": "Роль пользователя",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/groups": {
            "get": {
                "tags": [
                    "Group"
                ],
                "summary": "Список групп",
                "operationId": "dd307475c140d9d7218eda49e94ed6fb",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Количество записей",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Смещение",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total": {
                                            "description": "Общее количество записей",
                                            "type": "integer"
                                        },
                                        "groups": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Group"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Group"
                ],
                "summary": "Создать группу",
                "operationId": "16452bda6ae9c6c6a6db33cce1f5f27b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateGroupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/groups/{uuid}": {
            "put": {
                "tags": [
                    "Group"
                ],
                "summary": "Обновить группу",
                "operationId": "62008c5414d4de5a6773a5ae75e1ad4a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGroupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Group"
                ],
                "summary": "Удалить группу",
                "operationId": "763a8b718ef34f1d2b16d597448c73ee",
                "responses": {
                    "200": {
                        "description": "Ответ",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginRequest": {
                "description": "Тело запроса для назначения теста",
                "properties": {
                    "login": {
                        "description": "Логин",
                        "type": "string"
                    },
                    "password": {
                        "description": "Пароль",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CreateGroupRequest": {
                "description": "Тело запроса для создания группы",
                "properties": {
                    "name": {
                        "description": "Название группы",
                        "type": "string",
                        "example": "Frontend"
                    },
                    "employees": {
                        "type": "array",
                        "items": {
                            "description": "ID пользователя",
                            "type": "string",
                            "format": "uuid",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    }
                },
                "type": "object"
            },
            "UpdateGroupRequest": {
                "description": "Тело запроса для обновления группы",
                "properties": {
                    "name": {
                        "description": "Название группы",
                        "type": "string",
                        "example": "Frontend"
                    },
                    "employees": {
                        "type": "array",
                        "items": {
                            "description": "ID пользователя",
                            "type": "string",
                            "format": "uuid",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    }
                },
                "type": "object"
            },
...
            "Group": {
                "description": "Группа",
                "properties": {
                    "id": {
                        "description": "ID группы",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "Название группы",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}